functional component and changed folders in components
